site_stock/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
better-profanity==0.7.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-ckeditor-5==0.2.4
django-geojson==4.0.0
django-js-asset==2.0.0
django-leaflet==0.28.3
django-phonenumber-field==7.1.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
django-sslserver==0.22
django-timezone-field==5.0
dnspython==2.3.0
eventlet==0.33.3
flake8==6.0.0
flower==1.2.0
gevent==22.10.2
greenlet==2.0.2
humanize==4.6.0
idna==3.4
isort==5.12.0
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
redis==4.5.4
requests==2.29.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tomli==2.0.1
tornado==6.3.1
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
